+++ src/anschluss/pcf8574.rs
@@ -77,7 +77,7 @@ impl I2cMitPins {
 }
 
 impl I2cSettings {
-    fn aktiviert(&self, i2c_bus: I2cBus) -> bool {
+    pub(crate) fn aktiviert(&self, i2c_bus: I2cBus) -> bool {
         match i2c_bus {
             I2cBus::I2c0_1 => self.i2c0_1,
             // I2cBus::I2c2 => self.i2c2,
@@ -170,6 +170,7 @@ impl Lager {
 ///
 /// Anmerkung: Es ist möglich Software-I2C auf normalen Gpio-Pins zu aktivieren.
 /// Beachte dazu den Abschnitt "Aktivieren zusätzlicher I2C-Busse" in der `README.md`.
+///
 /// ACHTUNG: Dabei werden nur die Standard-Pins, die auch bei Pi4 verwendet werden, unterstützt.
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Serialize, Deserialize)]
 pub enum I2cBus {